+++ package.json
@@ -1,9 +1,11 @@
+
 {
   "dependencies": {
     "axios": "^0.19.0",
     "react-loader-spinner": "^3.1.2",
     "react-redux": "^7.1.0",
     "redux": "^4.0.4",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "styled-components": "^4.3.2"
   }
 }
