+++ react-redux-app/src/App.js
@@ -1,14 +1,15 @@
 import React from 'react';
-import './App.css';
 
-import ProfileList from './components/ProfileList';
+
 
 
 function App() {
   return (
-    <div className="App">
-      <ProfileList />
-    </div>
+        <div className="App">
+          <h1>Testing</h1>
+        </div>
+      
+    
   );
 }
