--- react-redux-app/src/actions/index.js
@@ -1,2 +1,0 @@
-export { getData, FETCH_PROFILE_DATA_START, FETCH_PROFILE_DATA_SUCCESS, FETCH_PROFILE_DATA_FAILURE } from './ProfileActions';
-
