+++ react-redux-app/src/components/reducers/index.js
@@ -1,0 +1,1 @@
+export { reducer } from './ImageReducer';
