+++ react-redux-app/src/index.js
@@ -8,7 +8,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 
-import { reducer } from './reducers';
+import { reducer } from './components/reducers';
 
 
 const store = createStore(reducer, applyMiddleware(thunk));
