--- react-redux-app/src/reducers/index.js
@@ -1,1 +1,0 @@
-export { reducer } from './ProfileReducer';
